Define Column.where to be used across DataFrame/Series
